Move non-templated functions in separate source
